# tests/EventBooking.Web.Tests/Contracts/openapi-v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "EventBooking API",
    "description": "EventBooking staff API and anonymous Attendee booking links.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5001/"
    }
  ],
  "paths": {
...
    "/api/events": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "List events.",
        "description": "Reads events filtered by location, date range and appointment type. A Manager sees their own type's capacity; an Admin or Coordinator sees every type.",
        "operationId": "listEvents",
        "parameters": [
          {
            "name": "locationId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "appointmentTypeId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOfEventResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ],
        "x-eventbooking-list": true,
        "x-mcp-tool": "list_events",
        "x-agent-hints": {
          "readOnly": true,
          "destructive": false,
          "idempotent": true,
          "openWorld": false
        }
      }
    },
    "/api/events/cancellable": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "List cancellable events.",
        "description": "Reads the active events whose window has not started, which are the ones a cancellation can still reach (FR-7.2).",
        "operationId": "listCancellableEvents",
        "parameters": [
          {
            "name": "locationId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOfEventResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ],
        "x-eventbooking-list": true,
        "x-capability": "ViewEventOperations",
        "x-mcp-tool": "list_cancellable_events",
        "x-agent-hints": {
          "readOnly": true,
          "destructive": false,
          "idempotent": true,
          "openWorld": false
        }
      }
    },
    "/api/events/{id}": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Read one event.",
        "description": "Reads one event with its capacities, filtered the same way the list is. An event outside the caller's scope reads as not found.",
        "operationId": "getEvent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ],
        "x-mcp-tool": "get_event",
        "x-agent-hints": {
          "readOnly": true,
          "destructive": false,
          "idempotent": true,
          "openWorld": false
        }
      }
    },
    "/api/events/{id}/capacities/{appointmentTypeId}": {
      "put": {
        "tags": [
          "Events"
        ],
        "summary": "Adjust an event capacity.",
        "description": "Replaces the total headcount for one appointment type on an active event. The type must be the caller's own, and the total must still cover every active booking.",
        "operationId": "adjustEventCapacity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "appointmentTypeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustCapacityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdjustEventCapacityOutcome"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ],
        "x-capability": "ManageEventNegotiation",
        "x-mcp-tool": "adjust_event_capacity",
        "x-agent-hints": {
          "readOnly": false,
          "destructive": true,
          "idempotent": true,
          "openWorld": false
        }
      }
    },
    "/api/events/{id}/cancel": {
      "post": {
        "tags": [
          "Events"
        ],
        "summary": "Cancel an event.",
        "description": "Cancels an event, voiding its bookings and re-inviting the affected attendees. Two-step: without confirm=true the call reports its consequence and changes nothing.",
        "operationId": "cancelEvent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "confirm",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelEventOutcome"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ],
        "x-capability": "CancelEvent",
        "x-mcp-tool": "cancel_event",
        "x-agent-hints": {
          "readOnly": false,
          "destructive": true,
          "idempotent": true,
          "openWorld": false
        }
      }
    },
...
    "/api/audit/events/{id}": {
      "get": {
        "tags": [
          "Audit"
        ],
        "summary": "Read an event's history.",
        "description": "Reads the audit history of one event, including its proposal.",
        "operationId": "getEventAuditHistory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOfAuditRowResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ],
        "x-eventbooking-list": true,
        "x-capability": "ViewEventAudit",
        "x-mcp-tool": "get_event_audit_history",
        "x-agent-hints": {
          "readOnly": true,
          "destructive": false,
          "idempotent": true,
          "openWorld": false
        }
      }
    },
    "/api/appointment-workspace/events": {
      "get": {
        "tags": [
          "Appointment Workspace"
        ],
        "summary": "List workspace events.",
        "description": "Reads the events listing the caller's type and ending between seven days ago and fourteen days ahead.",
        "operationId": "listWorkspaceEvents",
        "parameters": [
          {
            "name": "locationId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOfWorkspaceEventResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ],
        "x-eventbooking-list": true,
        "x-capability": "ConductAppointments",
        "x-mcp-tool": "list_workspace_events",
        "x-agent-hints": {
          "readOnly": true,
          "destructive": false,
          "idempotent": true,
          "openWorld": false
        }
      }
    },
    "/api/appointment-workspace/events/{eventId}": {
      "get": {
        "tags": [
          "Appointment Workspace"
        ],
        "summary": "Read a workspace roster.",
        "description": "Reads the minimum-data roster for one event (FR-8.1): name, email, the caller's own type's appointment status, and nothing else.",
        "operationId": "getWorkspaceRoster",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOfWorkspaceRosterRowResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ],
        "x-eventbooking-list": true,
        "x-capability": "ConductAppointments",
        "x-mcp-tool": "get_workspace_roster",
        "x-agent-hints": {
          "readOnly": true,
          "destructive": false,
          "idempotent": true,
          "openWorld": false
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdjustCapacityRequest": {
        "required": [
          "totalHeadcount"
        ],
        "type": "object",
        "properties": {
          "totalHeadcount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "AdjustEventCapacityOutcome": {
        "required": [
          "eventId",
          "totalHeadcount",
          "remainingCapacity",
          "changed"
        ],
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "description": "The event identifier.",
            "format": "uuid"
          },
          "totalHeadcount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "The total after the call.",
            "format": "int32"
          },
          "remainingCapacity": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "The remaining count after the call.",
            "format": "int32"
          },
          "changed": {
            "type": "boolean",
            "description": "Whether anything moved."
          }
        },
        "description": "The capacity row after adjusting."
      },
...
      "ApiLink": {
        "required": [
          "href",
          "method",
          "operationId"
        ],
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "operationId": {
            "type": "string"
          }
        }
      },
...
      "CancelEventOutcome": {
        "required": [
          "cancelledCount",
          "reinvitedCount",
          "awaitingAvailabilityCount"
        ],
        "type": "object",
        "properties": {
          "cancelledCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Bookings moved to cancelled (preview: bookings that would be).",
            "format": "int32"
          },
          "reinvitedCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Cancelled attendees issued a replacement invite.",
            "format": "int32"
          },
          "awaitingAvailabilityCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Cancelled attendees with no replacement.",
            "format": "int32"
          }
        },
        "description": "How many bookings were cancelled, rebooked, or left awaiting availability."
      },
...
      "EventCapacityResponse": {
        "required": [
          "appointmentTypeId",
          "code",
          "name",
          "totalHeadcount",
          "remainingCapacity",
          "_links"
        ],
        "type": "object",
        "properties": {
          "appointmentTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "totalHeadcount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "remainingCapacity": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "_links": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/ApiLink"
            }
          }
        }
      },
...
      "EventResponse": {
        "required": [
          "id",
          "proposalId",
          "locationId",
          "locationCode",
          "locationName",
          "time",
          "status",
          "capacities",
          "activeBookings",
          "_links"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "proposalId": {
            "type": "string",
            "format": "uuid"
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "locationCode": {
            "type": "string"
          },
          "locationName": {
            "type": "string"
          },
          "time": {
            "$ref": "#/components/schemas/EventTimeResponse"
          },
          "status": {
            "type": "string"
          },
          "capacities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventCapacityResponse"
            }
          },
          "activeBookings": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "_links": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/ApiLink"
            }
          }
        }
      },
      "EventTimeResponse": {
        "required": [
          "date",
          "startTime",
          "durationMinutes",
          "startLocal",
          "endLocal",
          "startUtc",
          "endUtc",
          "timeZoneId",
          "zoneAbbreviation"
        ],
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "startTime": {
            "type": "string",
            "format": "time"
          },
          "durationMinutes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "startLocal": {
            "type": "string",
            "format": "date-time"
          },
          "endLocal": {
            "type": "string",
            "format": "date-time"
          },
          "startUtc": {
            "type": "string",
            "format": "date-time"
          },
          "endUtc": {
            "type": "string",
            "format": "date-time"
          },
          "timeZoneId": {
            "type": "string"
          },
          "zoneAbbreviation": {
            "type": "string"
          }
        }
      },
...
      "PageOfAuditRowResponse": {
        "required": [
          "items",
          "nextCursor"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditRowResponse"
            }
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PageOfEventResponse": {
        "required": [
          "items",
          "nextCursor"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventResponse"
            }
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PageOfWorkspaceEventResponse": {
        "required": [
          "items",
          "nextCursor"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkspaceEventResponse"
            }
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "PageOfWorkspaceRosterRowResponse": {
        "required": [
          "items",
          "nextCursor"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkspaceRosterRowResponse"
            }
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}